ci: harden GitHub Actions pipelines (#23)
* ci: harden github actions pipelines

* ci: fix security workflow failures

* ci: avoid unsupported code scanning uploads

* ci: tighten workflow security gates

* ci: harden publish workflow permissions

* ci: allow codeql workflow metadata access

* ci: run codeql without code scanning upload

+++ crates/core/src/workspace.rs
@@ -1150,7 +1150,7 @@ impl WorkspaceStore {
             process.pid > 0,
             "terminal process {process_id} has invalid pid"
         );
-        stop_process(process.pid as u32)?;
+        stop_process(process.pid)?;
         self.mark_terminal_process_stopped(process.id, Some(SIGTERM_EXIT_CODE))
     }
